# openapi.yaml
openapi: 3.0.0
info:
  title: Main Collection
  version: 1.0.0
  description: Contains the Postman docs for all of our api/endpoionts.
servers:
  - url: http://localhost:3000
paths:
...
  /api/admin/ban/{userId}:
    patch:
      summary: Ban User
      responses: {}
      tags:
        - Admin
      security:
        - BearerAuth: []
      description: >-
        # **PATCH /api/admin/ban/\[userId\]**


        Allows an **admin** to **ban or unban a user** by updating their
        `isBanned` status.


        Authentication is required, and only users with the **ADMIN** role can
        perform this action.


        ---


        # **Authentication**


        Include a **Bearer token** in the `Authorization` header:


        ```

        Authorization: Bearer <access_token>

         ```

        The token must be a valid JWT for a user with `role = ADMIN`.


        ---


        # **Path Parameters**


        | Parameter | Type | Required | Description |

        | --- | --- | --- | --- |

        | `userId` | number | Yes | The ID of the user to ban or unban |


        ---


        # **Request Body**


        | Field | Type | Required | Description |

        | --- | --- | --- | --- |

        | `result` | string | Yes | Must be either `"BAN"` to ban the user or
        `"UNBAN"` to unban the user |


        ---


        ## **Example Request**


        ```

        PATCH /api/admin/ban/5

         ```

        Header:


        ```

        Authorization: Bearer <admin_access_token>

         ```

        Body:


        ```

        {  "result": "BAN"}

         ```

        ---


        # **Responses**


        ## **200 OK — User Ban Status Updated**


        Returned when the user is successfully banned or unbanned.


        ```

        {  
             "id": 5,  
             "username": "exampleUser",  
             "email": "user@example.com",  
             "role": "USER",  
             "isBanned": true,  
             "createdAt": "2026-03-01T12:00:00.000Z",  
             "updatedAt": "2026-03-06T16:00:00.000Z"
        }

         ```

        ---


        ## **400 Bad Request**


        Returned if:


        - The `result` field is missing or invalid
            
        - The `userId` parameter is missing or not a number
            

        Examples:


        ```

        {  "error": "Ban result is required"}

         ```

        ```

        {  "error": "Ban result must be a string"}

         ```

        ```

        {  "error": "Invalid ban result"}

         ```

        ```

        {  "error": "Valid userId parameter is required"}

         ```

        ---


        ## **401 Unauthorized**


        Returned if:


        - The request does not include a valid token
            
        - The token is invalid or expired
            
        - The admin user does not exist
            

        ```

        {  "error": "Unauthorized"}

         ```

        Example:


        ```

        {  "error": "Invalid token"}

         ```

        ---


        ## **403 Forbidden**


        Returned if the authenticated user is **not an admin**.


        ```

        {  "error": "Forbidden"}

         ```

        ---


        ## **404 Not Found**


        Returned if the target user does not exist.


        ```

        {  "error": "User not found"}

         ```

        ---


        ## **500 Internal Server Error**


        Returned if an unexpected server error occurs.


        ```

        {  "error": "Internal Server Error"}

         ```

        ---


        # **Notes**


        - Only **admin users** can ban or unban other users.
            
        - The `result` field must be `"BAN"` or `"UNBAN"` (case-insensitive).
            
        - The response returns the **updated user object**, including `isBanned`
        status.
            
        - This endpoint does **not delete the user account**—it only updates the
        `isBanned` flag.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                result:
                  type: string
            example:
              result: UNBAN
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []